# backend/docs/api/phone-store-openapi.yaml
openapi: 3.0.3
info:
  title: Phone Store E-Commerce Platform API
  version: 1.0.0
  description: API documentation for the Phone Store E-Commerce Platform backend services.
servers:
  - url: http://localhost:8080
    description: local dev server
...
paths:
  /api/v1/auth/email-verifications:
    post:
      tags:
        - Authentication
      summary: API Endpoint
      operationId: resendVerificationEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequest'
      responses:
        '202':
          description: API Description
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/email-verifications/confirm:
    post:
      tags:
        - Authentication
      summary: API Endpoint
      operationId: confirmEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationConfirmRequest'
      responses:
        '204':
          description: API Description
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: API Endpoint
      description: API Description
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                InvalidCredentials:
                  summary: API Endpoint
                  value:
                    errorCode: BAD_REQUEST
                    message: Invalid username or password
                    correlationId: a6b7c8d9-e0f1-42a3-b4c5-d6e7f8a9b0c1
                UserDeactivated:
                  summary: API Endpoint
                  value:
                    errorCode: BAD_REQUEST
                    message: User account is deactivated
                    correlationId: 2df3e485-bc75-4d0f-8973-455bdaef3841
                ValidationFailed:
                  summary: API Endpoint
                  value:
                    errorCode: VALIDATION_FAILED
                    message: Error
                    fieldErrors:
                      username: Username is required
                      password: Password is required
                    correlationId: dcba0987-7654-3210-fedc-ba9876543210
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: API Endpoint
      operationId: logoutUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: API Description
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/password-reset-requests:
    post:
      tags:
        - Authentication
      summary: API Endpoint
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '202':
          description: API Description
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/password-resets/confirm:
    post:
      tags:
        - Authentication
      summary: API Endpoint
      operationId: confirmPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '204':
          description: API Description
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: API Endpoint
      description: API Description
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                UsernameTaken:
                  summary: API Endpoint
                  value:
                    errorCode: BAD_REQUEST
                    message: Username is already taken
                    correlationId: df185ee6-d7e7-494b-a63e-f1b203c9bf82
                EmailRegistered:
                  summary: API Endpoint
                  value:
                    errorCode: BAD_REQUEST
                    message: Email is already registered
                    correlationId: a5f2da51-419b-4e08-9fd1-cb4be6c1e345
                ValidationFailed:
                  summary: API Endpoint
                  value:
                    errorCode: VALIDATION_FAILED
                    message: Error
                    fieldErrors:
                      username: Username must be between 3 and 100 characters
                      email: Invalid email format
                      password: Password must be between 8 and 64 characters
                    correlationId: 3d50b4f8-e1a1-432d-8b01-df5f1d93bff0
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/token/refresh:
    post:
      tags:
        - Authentication
      summary: API Endpoint
      description: API Description
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/me:
    get:
      tags:
        - Profile
      summary: API Endpoint
      operationId: getCurrentProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Profile
      summary: API Endpoint
      operationId: updateCurrentProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
      responses:
        '200':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/admin/coupons:
    post:
      tags:
        - Admin Coupons
      summary: Create Coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Admin Coupons
      summary: get all coupons
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/coupons/{code}:
    get:
      tags:
        - Admin Coupons
      summary: Get Coupon
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Admin Coupons
      summary: Invalidate Coupon
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/coupons/{code}/targets:
    post:
      tags:
        - Admin Coupons
      summary: Add Subjects
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    BadRequest:
      description: L?i d? li?u d?u vào (Thi?u ho?c sai d?nh d?ng tham s?)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            MissingParameter:
              summary: Thi?u tham s?
              value:
                errorCode: BAD_REQUEST
                message: Tham s? b?t bu?c không du?c d? tr?ng
                correlationId: e3b0c442-uuid-abc
            InvalidFormat:
              summary: Sai d?nh d?ng UUID/Email
              value:
                errorCode: BAD_REQUEST
                message: Ð?nh d?ng email không h?p l?
                correlationId: e3b0c442-uuid-def
    Unauthorized:
      description: K? m?o danh (Chua dang nh?p ho?c Token h?t h?n)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            TokenExpired:
              summary: JWT h?t h?n
              value:
                errorCode: UNAUTHORIZED
                message: Access token dã h?t h?n
                correlationId: e3b0c442-uuid-123
    Forbidden:
      description: B? t? ch?i (Không d? quy?n h?n truy c?p chuyên sâu)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            AccessDenied:
              summary: Vu?t quy?n
              value:
                errorCode: FORBIDDEN
                message: Tài kho?n c?a b?n không có d?c quy?n (Admin/Manager) d? g?i API này
                correlationId: e3b0c442-uuid-789
    NotFound:
      description: Không tìm th?y (Tài nguyên không t?n t?i trên DB)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            ResourceNotFound:
              summary: D? li?u tr?ng
              value:
                errorCode: RESOURCE_NOT_FOUND
                message: Không tìm th?y d? li?u id=9999 (Entity Not Found)
                correlationId: e3b0c442-uuid-444
    UnprocessableEntity:
      description: L?i logic nghi?p v? (Ki?m tra constraint/business th?t b?i)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            BusinessConflict:
              summary: H?t t?n kho
              value:
                errorCode: BUSINESS_CONFLICT
                message: S? lu?ng t?n kho không d? d? th?c hi?n yêu c?u!
                correlationId: e3b0c442-uuid-555
            ValidationFailed:
              summary: Vi ph?m constraint
              value:
                errorCode: VALIDATION_FAILED
                message: Validation Error
                fieldErrors:
                  quantity: S? lu?ng ph?i l?n hon 0
                  price: Giá bán không du?c âm
                correlationId: e3b0c442-uuid-666
    InternalServerError:
      description: L?i máy ch? c?c b? (Exception backend không mong mu?n)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            ServerError:
              summary: Null Pointer / Database Down
              value:
                errorCode: INTERNAL_SERVER_ERROR
                message: Ðã x?y ra l?i h? th?ng nghiêm tr?ng, vui lòng liên h? Admin
                correlationId: e3b0c442-uuid-999
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
          example: newuser
        email:
          type: string
          format: email
          maxLength: 255
          example: newuser@example.com
        password:
          type: string
          minLength: 8
          maxLength: 64
          example: password123
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: loginuser
        password:
          type: string
          example: password123
    AuthResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - tokenType
        - username
        - role
      properties:
        accessToken:
          type: string
          description: API Description
          example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJsb2dpbnVzZXIiLCJyb2xlIjoiVVNFUiIsImlhdCI6MTc...
        refreshToken:
          type: string
          description: API Description
          example: bda3e8c9-fa20-4b2a-87ef...
        tokenType:
          type: string
          example: Bearer
        username:
          type: string
          example: loginuser
        role:
          type: string
          example: USER
...
    ApiError:
      type: object
      required:
        - errorCode
        - message
        - correlationId
      properties:
        errorCode:
          type: string
          description: API Description
          example: BAD_REQUEST
        message:
          type: string
          description: API Description
          example: Invalid username or password
        correlationId:
          type: string
          format: uuid
          description: API Description
          example: a6b7c8d9-e0f1-42a3-b4c5-d6e7f8a9b0c1
        fieldErrors:
          type: object
          additionalProperties:
            type: string
          description: API Description
          example:
            username: Username is required
            password: Password is required
...
    TokenRefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    EmailVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    EmailVerificationConfirmRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    PasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    PasswordResetConfirmRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
        newPassword:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
        accountStatus:
          type: string
        emailVerifiedAt:
          type: string
          format: date-time
          nullable: true
    UserProfileUpdateRequest:
      type: object
      properties:
        phone:
          type: string
        avatarUrl:
          type: string
...